Add files via upload
The translation API provides an interface for translation. You can also detect the language of the text with this python language translation video.

You will learn how to list available languages and translate them to another language with the help of this Python google translate tutorial.

If you know basic Python, then this project is for you. Practice with it and get better at programming. It can translate text immediately, and you can get to know about the language from pasting it in the search space.

Enabling translation API and authentication API requests are also discussed in this video; our trainer will explain it all from scratch with proper examples. Installing the Python client library, translating text, and detecting language are also discussed in detail.